migrated recontruct_areacello to upstream_stash
